[ET][Portable] Introduce advanced index util
Refactor `index.Tensor` helper functions into `advanced_index_util`, to be reused by `index_put`

Differential Revision: [D50031047](https://our.internmc.facebook.com/intern/diff/D50031047/)

[ghstack-poisoned]
